requirements.txt
streamlit==1.45.0
torch==2.7.0
torchvision==0.22.0
Pillow==11.2.1
numpy==2.2.5
opencv-python==4.11.0.86
matplotlib==3.10.3 